Move files around into sensible directories
