+++ jigsaw/client/__init__.py
@@ -1,0 +1,2 @@
+#!/usr/bin/venv python3
+from .gui import GameClient
