+++ tox.ini
@@ -21,6 +21,9 @@ ignore=
     # Comments
     E266
 
+per-file-ignores =
+    jigsaw/client/__init__.py:F401
+
 [isort]
 # Select the 5th style (Hanging grid grouped) to handle longer import.
 # This choice is mostly arbitrary and can be changed at your will.
